+++ modules/server/src/internal/Server.ts
@@ -7,7 +7,7 @@ import * as isPromise from 'is-promise';
 import { IController, Controller } from '../Controller';
 import { IInjector } from '../Injector';
 import { Status } from '../enums';
-import { IInternalRoute, IInternalInjectedRoute } from '../interfaces';
+import { IInternalRoute, IInternalInjectedRoute, IServerConfig } from '../interfaces';
 import { Result, JsonResult, AssetResult } from '../results';
 import { ILogger, ConsoleLogger } from '../loggers';
 import { IncomingForm } from 'formidable';
@@ -31,9 +31,9 @@ export class InternalServer {
      */
     private _dependency: Dependency;
     /**
-     * The locations of the static folders based on their routes and folder locations.
+     * Server Configuration to change the flow of the server.
      */
-    private _staticFolders: string[][];
+    private _config: IServerConfig;
     /**
      * The routes that should be used be the server to determine which controller and method should be used.
      */
@@ -51,17 +51,9 @@ export class InternalServer {
      */
     private _logger: ILogger;
     /**
-     * The current directory that the server is running in.
-     */
-    private _dir: string;
-    /**
      * The list of middleware callbacks that need to be included before the found injected routes and route.
      */
     private _middlewareCallbacks: ((req: http.IncomingMessage | http2.Http2ServerRequest, res: http.ServerResponse | http2.Http2ServerResponse, next: (err?: any) => void) => void)[];
-    /**
-     * The upload directory that should be used for formidable.
-     */
-    public uploadDir: string;
 
     /**
      * Constructor method that will build out a basic internal server.
@@ -70,16 +62,14 @@ export class InternalServer {
      * @param dir The directory that the server should be running in.
      * @param logger The custom logger that should be used to get information as the server runs.
      */
-    constructor(dependency: Dependency, dir?: string, logger?: ILogger) {
-        this._dir = dir || '';
+    constructor(dependency: Dependency, config: IServerConfig, logger?: ILogger) {
+        this._config = config;
         this._middlewareCallbacks = [];
         this._dependency = dependency;
         this._faviconRegex = /favicon\.icon$/;
         this._injectedRoutes = new RouteContainer<IInternalInjectedRoute>();
         this._routes = new RouteContainer<IInternalRoute>();
-        this._staticFolders = [];
         this._logger = logger ? logger : new ConsoleLogger();
-        this.uploadDir = '';
     }
 
     /**
@@ -110,15 +100,6 @@ export class InternalServer {
                 route.injector = injectors[i];
             }
         }
-    }
-
-    /**
-     * Method is meant to register static folder locations.
-     * 
-     * @param folders The static folders that need to be added to the server.
-     */
-    public registerStaticLocations(...folders: string[]) {
-        this._staticFolders = this._staticFolders.concat(folders.map(x => x.split('/')));
     }
 
     /**
@@ -193,8 +174,8 @@ export class InternalServer {
      */
     private getStaticResult(parsedUrl: url.UrlWithParsedQuery): Result {
         let currentPath = parsedUrl.pathname === undefined || parsedUrl.pathname === null ? '' : parsedUrl.pathname;
-        let fullPath = path.join(this._dir, currentPath);
-        return new AssetResult(fullPath);
+        let fullPath = path.join(this._config.cwd, currentPath);
+        return new AssetResult(fullPath, true);
     }
 
     /**
@@ -204,12 +185,17 @@ export class InternalServer {
      */
     private isStaticResource(parsedUrl: url.UrlWithParsedQuery) {
         let splitPath = parsedUrl.pathname === undefined || parsedUrl.pathname === null ? [] : parsedUrl.pathname.split('/');
-        for (let i = 0; i < this._staticFolders.length; ++i) {
-            let isStaticResource = true;
-            for (let j = 0; j < this._staticFolders[i].length && isStaticResource; ++j) {
-                isStaticResource = isStaticResource && this._staticFolders[i][j] === splitPath[j];
+        let isStaticResource = true;
+        if (this._config.staticFolders !== undefined && this._config.staticFolders.length > 0) {
+            for (let i = 0; i < this._config.staticFolders.length; ++i) {
+                let splitStatic = this._config.staticFolders[i].split('/');
+
+                let isStaticResource = true;
+                for (let j = 0; j < splitStatic.length && isStaticResource; ++j) {
+                    isStaticResource = isStaticResource && splitStatic[j] === splitPath[j];
+                }
+                if (isStaticResource) return true;
             }
-            if (isStaticResource) return true;
         }
         return false;
     }
@@ -239,8 +225,8 @@ export class InternalServer {
             if (this.isStaticResource(parsedUrl)) {
                 response = this.getStaticResult(parsedUrl);
             } else if (this.isFavicon(parsedUrl)) {
-                let fullPath = path.join(this._dir, 'favicon.ico');
-                response = new AssetResult(fullPath);
+                let fullPath = path.join(this._config.cwd, 'favicon.ico');
+                response = new AssetResult(fullPath, true);
             } else {
                 let routes = this._routes.find(parsedUrl, req.method || '');
                 let injectors = this._injectedRoutes.find(parsedUrl, req.method || '');
@@ -266,7 +252,7 @@ export class InternalServer {
             response.body = { message: 'Internal Server Error' };
         } finally {
             try {
-                await response.processResponse(res);
+                await response.processResponse(res, this._config);
             } catch (err) {
                 this._logger.error(err);
                 this.sendError(req, res);
@@ -338,7 +324,7 @@ export class InternalServer {
             this.processBeforeRoute(route, dependency)
                 .then(x => {
                     let controller = <any>dependency.resolve(route.controller);
-                    controller._dirname = this._dir;
+                    controller._dirname = this._config.cwd;
 
                     if (x !== undefined) return resolve(x);
                     let response = controller[route.key].apply(controller, this.processArguments(route, body, parsedUrl, controller));
@@ -412,8 +398,8 @@ export class InternalServer {
         return new Promise<IFormModel>((resolve, reject) => {
             let form = new IncomingForm();
             let result: any = {};
-            if (this.uploadDir.length > 0) {
-                form.uploadDir = this.uploadDir;
+            if (this._config.uploadDir !== undefined && this._config.uploadDir.length > 0) {
+                form.uploadDir = this._config.uploadDir;
             }
             form.parse(req, (err, fields, files) => {
                 if (err) return reject(err);
@@ -464,7 +450,7 @@ export class InternalServer {
         let response = new JsonResult();
         response.status = Status.InternalServerError;
         response.body = { message: 'Internal Server Error' };
-        await response.processResponse(res);
+        await response.processResponse(res, this._config);
     }
     //#endregion
     //#endregion