+++ modules/server/src/results/fileResult.ts
@@ -1,6 +1,7 @@
 import { Result } from './result';
 import { Http2ServerResponse } from 'http2';
 import { ServerResponse } from 'http';
+import { IServerConfig } from '../interfaces';
 import * as fs from 'fs';
 
 /**
@@ -39,10 +40,10 @@ export class FileResult extends Result {
      * 
      * @param res The server response object that we need to work with when processing this result.
      */
-    public async processResponse(res: Http2ServerResponse | ServerResponse) {
+    public async processResponse(res: Http2ServerResponse | ServerResponse, config: IServerConfig) {
         this.headers['Content-Type'] = this._contentType;
         this.headers['Content-Disposition'] = `attachment; filename=${this._fileName}`;
         this.headers['Content-Length'] = (<Buffer>this.body).length.toString();
-        super.processResponse(res);
+        super.processResponse(res, config);
     }
 }