+++ modules/server/src/results/jsonResult.ts
@@ -1,6 +1,7 @@
 import { Result } from './result';
 import { Http2ServerResponse } from 'http2';
 import { ServerResponse } from 'http';
+import { IServerConfig } from '../interfaces';
 
 /**
  * A result that handles a basic object and stringifies it to send to the client.
@@ -11,9 +12,9 @@ export class JsonResult extends Result {
      * 
      * @param res The server response object that we need to work with when processing this result.
      */
-    public async processResponse(res: Http2ServerResponse | ServerResponse) {
+    public async processResponse(res: Http2ServerResponse | ServerResponse, config: IServerConfig) {
         this.headers['Content-Type'] = 'application/json';
         this.body = this.body === undefined ? 'null' :  JSON.stringify(this.body);
-        super.processResponse(res);
+        super.processResponse(res, config);
     }
 }